This library is about testing validation, not marshalling. So this changed the benchmark to the validation method only for Marshal.
Benchmarking in Marshal validation + marshalling next to libraries that are only benchmarked with validation is not a correct benchmark.
The new ts-json-validator is able to do both as well, but only validation is benchmarked. Thus using only the validation method in Marshal is the only way to make a fair comparison.